+++ spec/ey_yml_customized_deploy_spec.rb
@@ -47,10 +47,12 @@ describe "Deploying an app with ey.yml" do
       @deployer.enable_maintenance_page
       deploy_dir.join('shared','system','maintenance.html').should exist
       redeploy_test_application
-      read_output.should =~ /Maintenance page is still up. You must remove it manually./
+      read_output.should =~ /Maintenance page is still up./
       deploy_dir.join('shared','system','maintenance.html').should exist
       deploy_dir.join('current','maintenance_disabled').should_not exist
       deploy_dir.join('current','maintenance_enabled').should exist
+      @deployer.disable_maintenance_page
+      deploy_dir.join('shared','system','maintenance.html').should_not exist
     end
 
     it "makes custom variables available to hooks" do