+++ ChangeLog.md
@@ -15,6 +15,7 @@
   * `precompile_assets: true or false (default: inferred using app/assets and config/application.rb)`
   * Supports new ey.yml option to ignore the missing database adapter warning:
   * `ignore_database_adapter_warning: true (default: false)`
+  * Give better feedback when maintenance is enabled/disabled (or not).
 
 ## v1.6.4 (2012-04-26)